Fixed bug in splitGroups, tidied up
